Fix wrong assumption about the tablet pad strip position
This was a mistake when implementing support for tablet pad strips: the
position isn't integer! Or at least, not in the way we receive it.

libinput returns a clamped double with a range of [0, 1], but we
accidentally (and implicitly) converted it to an integer. And then,
Wayland clients assumedly only get either a 0 or a 1. This gets worse as
the Wayland protocol says the position is a range of [0, 65535].
Basically this never should have worked.

I fixed this by storing the intermediate double as a qreal in
TabletStripEvent, and then multiplying it by 65355 inside of the Wayland
interface like we do for other doubles.


(cherry picked from commit 1beb408098d2231f827fa584d231339d4bf44399)

+++ src/effect/effect.cpp
@@ -510,7 +510,7 @@ bool Effect::tabletPadButtonEvent(uint button, bool pressed, void *device)
     return false;
 }
 
-bool Effect::tabletPadStripEvent(int number, int position, bool isFinger, void *device)
+bool Effect::tabletPadStripEvent(int number, qreal position, bool isFinger, void *device)
 {
     return false;
 }
