+++ src/tablet_input.cpp
@@ -420,7 +420,7 @@ void KWin::TabletInputRedirection::tabletPadButtonEvent(uint button, bool isPres
     }
 }
 
-void KWin::TabletInputRedirection::tabletPadStripEvent(int number, int position, bool isFinger, quint32 group, quint32 mode, std::chrono::microseconds time, InputDevice *device)
+void KWin::TabletInputRedirection::tabletPadStripEvent(int number, qreal position, bool isFinger, quint32 group, quint32 mode, std::chrono::microseconds time, InputDevice *device)
 {
     TabletPadStripEvent event{
         .device = device,
